Add tests to ensure new tab button enabled for Lens error url
Lens Web has an error state when a payload parameter exists but no
payload is present. We want to make sure the Lens new tab button is
enabled so users can retry in a new tab if they please. This change add
tests to ensure that functionality.

Bug: b:245353468
Change-Id: Id13cc5047c815f50a0289e68b2961a298c7aace4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3954385
Commit-Queue: Duncan Mercer <mercerd@google.com>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Anudeep Palanki <apalanki@google.com>
Cr-Commit-Position: refs/heads/main@{#1059282}