Select-to-Speak: Expose option for voice switching
This change adds an option to the Select-to-Speak options page that
allows the user to toggle voice switching. The option is hidden
behind a new feature flag. We also gate the voice switching logic
behind the new flag, instead of the language detection flag. This will
allow us to launch voice switching without launching language
detection.

Note: It seems that automatic voice switching isn't fully implemented,
so a follow-up change will be added to complete this feature.

AX-Relnotes: N/A
Bug: 950391
Change-Id: I90ff692576b1d46ec28eb415871e4155ae065b5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3947402
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059288}