os settings: compile shared ts files
There are a set of ts files that's shared between os settings and
browser settings app.  Before this CL they are compiled with browser
settings build step and then copied as raw js files over to os settings.

This CL improves this a bit by copying the shared ts files as source ts
files and compile them with the rest of os settings.

This makes the build a little bit cleaner and enforces checks on the
shared files.  This CL already discovered a couple missing files that
was not included before.

Bug: 1315757
Change-Id: I34cbdc0c8e940441a70ed6831339608ecd3541e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3950337
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Commit-Queue: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059299}