Add feature WebRtcCombinedNetworkAndWorkerThread
This feature combines the WebRTC worker thread and the WebRTC network
threads onto the same thread. The worker thread is used for both in the
case the feature is on as the network thread has restrictions on blocking.

Tested: Meet meeting and in-tab peer-connection demo with both sides of the feature. Confirmed thread name in chrome://tracing.
Bug: 1373439
Change-Id: I80b4f072a287269d00ee44d879144ec30ac2272d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3945503
Reviewed-by: Markus Handell <handellm@google.com>
Commit-Queue: Evan Shrubsole <eshr@google.com>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: Tomas Gunnarsson <tommi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059302}