[GridFragmentation] Aggregate OOFs before PlaceOutOfFlowItems()
Instead of looping over all items to position the OOF grid children,
aggregate a list of OOFs when we loop over the grid items earlier on.
Use this list within PlaceOutOfFlowItems(), and store this list on the
grid break token to ensure we can reuse these between fragments.

This change should result in no functional changes - it is in
preparation for making a similar change as CL:3927797 for grid.

Bug: 1367912
Change-Id: Ifcece6ef04637ec6f2bd2b3721743bdb1aea1e5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3946350
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1059305}