Rename LaunchMode{,Recorder} in preparation for refactor
This CL prepares for a new launch mode metric, while keeping the old
launch mode around for a couple releases. It renames
LaunchModeRecorder to OldLaunchModeRecorder and LaunchMode to
OldLaunchMode. This will make it easier to remove later.
We will introduce a new LaunchMode enum and
LaunchModeRecorder with new values and logic, and a new histogram,
Launch.Modes2. We will remove the Old launch mode stuff after the new
launch has been in prod for a couple milestones.

Bug: 1366137
Change-Id: I2f0accbf6be4db577db98a4b26ad8c4d4fa74c4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3935224
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059351}