[GridFragmentation] OOF static position when CB is outside multicol
The static position of OOFs in a fragmented grid container were
determined within the last fragment. If the OOF's CB lives outside the
associated multicol, then that static position would be relative to
the last parent fragment and would become negative.

Similar to CL:3927797, attempt to determine the static position in
the fragment it belongs. Note that this may end up being incorrect
for center aligned items, or items with a grid-area that does not
start in the first or last fragment.

Bug: 1367912
Change-Id: I61bf29ccf97d2bb293ba7916f327883f43e53214
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3949467
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1059352}