[Read Anything] Check validity of model indices
Instead of setting the font name and line/letter spacing index values
directly from prefs, we first check if those values are valid and if so
we set those pref values as the default values.

Then, we set the initial values to be the default values. That way, if
the pref values were not valid, the original default values will be
used instead.

Bug: 1266555
Change-Id: If98668cb05fe1c72fcc6bf3c92941204018cd4de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3943748
Commit-Queue: Jocelyn Tran <jocelyntran@google.com>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/main@{#1059357}