Hide microphone mute toast for the privacy hub toggle
Previously, microphone mute toast was displayed only when the system
input mute state was toggled by the keyboard switch or when the hw
switch was toggled. Privacy hub is introducing a new button in the UI
that can be used to toggle the system input mute state. We will not
display the microphone mute toast if the system input mute state is
toggled by the privacy hub switch. It will already be very obvious for
the user from the UI that the system input mute has changed. Displaying
a toast in this scenario can seem redundant.

Bug: b/253221902
Change-Id: I88cbac141773c3c39dbb675f23cc6409f530ffd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3948566
Commit-Queue: Md Shahadat Hossain Shahin <shahinmd@google.com>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059363}