[Web Payment] Add a test for canmakepayment event fields
This patch adds a test that verifies feature detection of
"canmakepayment" event fields in a payment handler with and without the
chrome://flags/#clear-identity-in-can-make-payment flag. This test will
be useful when an upcoming patch is supposed to make these fields
"undefined" instead of empty.

Bug: 1290492
Change-Id: I1406d7ac9d2e627c203d26d6b92140f874d6bb81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3954282
Reviewed-by: Nick Burris <nburris@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059370}