Reland: [Toolbar] Post capture request when controls unlock
The event that triggers this capture is already relatively long,
so performing it synchronously risks dropping frames and/or delaying
the processing of a touch event, leading to the very jank we're
trying to prevent.

Reland notes: this makes it possible for a task to try to draw the
toolbar post-destroy, which we need to prevent.

Bug: 1063346
Change-Id: Ic8021a3416732369c1afd1fd1b3b86e9becf04ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3957275
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059425}