shelf: Fix ShelfPartyTest.PartyingItemsHiddenFromShelf
The test enables Shelf Party and waits for all animations to complete,
but Shelf Party animations can go on forever. It is a test of hiding
partying items from the shelf, and it made sense before the actual
partying item animations were added. Now it will wait one second instead
of waiting for all animations to finish.

To facilitate this fix, I am moving the test from LtrRtlShelfViewTest to
ShelfPartyTest because ShelfPartyTest uses
base::test::TaskEnvironment::TimeSource::MOCK_TIME. This also means that
the test will run with various shelf alignment and autohide behavior
instead of with left-to-right and right-to-left locale. That's okay!

Fixed: 1372295
Test: ash_unittests All/ShelfPartyTest.PartyingItemsHiddenFromShelf/*
Change-Id: I1713d3a62b136368072bdfb131faf2d643bb95a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3957278
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059441}