Provide BufferQueue mailbox for overlay testing the primary plane
When BufferQueue was enabled on Ash and Lacros overlays were broken
because this wasn't implemented. It was trying to test overlays with a
zero-mailbox and failing. This will fix that.

This was not a problem on Android (where BufferQueue is still enabled)
because it doesn't do overlay testing.

I've confirmed this fix locally on a Volteer and an Octopus device. Normal overlays are promoted, and PIP rounded corner overlays are
promoted as well, with and without BufferQueue enabled.

Enabling BufferQueue on Ash and Lacros will be done in separate CLs in case they require a rollback.

Change-Id: I5b8e8f72e6b3401b90e3997ba75c939788cd0e93
Bug: 1322528,1364553
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3949357
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Commit-Queue: Kevin Haslett <khaslett@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059450}