Remove unused status area padding
Before Calendar was enabled, there was a padding that used to separate the status area clock from other tray items to make it distinguishable.

Since the time is now displayed next to the calendar, there is no need to add this extra padding. This padding was set to zero if the Calendar flag was enabled, so now that Calendar has reached stable there won't be any visible changes after this clean-up.

Bug: 1338748
Change-Id: Ibf1195ff9334c144686e68c469faef4c103e56b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645194
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Kevin Radtke <kradtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059457}