Reduce runtime of All/VideoPaintPreviewTest blink_unittests.
These tests took 29s to run, reduce this to 335ms.

These tests were waiting on loading events when creating the media
player. Use the StubWebMediaPlayer which is used in the other tests in
this file and that causes the Load to happen immediately instead of
timing out.

Change-Id: Ie915d4b1d09308f52e608a797a316cb6d6db9a19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3953819
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059472}