[dPWA] Moving reusable functions from manifest update task to utils
This CL introduces a manifest_update_utils file that stores some
enums and utility functions that will be used by the ManifestUpdateManager,
ManifestUpdateTask and the 2 commands. This is broken from
crrev.com/c/3934704 which will be rebased on top of this.

This leads to some duplicate functions created inside ManifestUpdateTask
which will be removed once the task is deleted.

Bug: 1298130, b/243699195
Change-Id: Ib7126e146615b1f7a9d89594620dba18b5623170
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3953580
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Phillis Tang <phillis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059491}