Fix parent creation when building the scroll tree.
PropertyTreeManager::CreateCompositorScrollNode expects that we have
already created the scroll parent, but this may fail for a scroller in a
fixed position container in a subframe.

There is existing code to address this case but it did not run with
unification.  The comment that unification "will already have done this"
is not correct, because this is the same code that unification uses to
set up the scroll tree (via EnsureCompositorScrollNodes).

Bug: 1374044, b/253298018
Change-Id: I0b63752a80257b76bfe0f2b4b192c784daabae18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3957535
Commit-Queue: Steve Kobes <skobes@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059493}