wm: Prevent destroying the backdrop while hiding it
It was possible that `BackdropController::Hide()` calls itself
recursively indirectly. As a result, hiding the backdrop may
lead to updating the backdrop state when there's no window
needing a backdrop, resulting in `Hide()` getting called again
to delete the backdrop while still in the stack of the previous
`Hide()` call. This can lead to a UAF.

This CL fixes this issue, and adds a regression test that crashes
with ASAN turned on.

Bug: 1368587
Test: Manually. Added a new test.
Change-Id: I4f55d3595eb06e7bc296dd45c4f2b3240488ef7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3953903
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059497}