[omnibox][post-ac] Move stability/latency metrics to `NotifyChanged()`.
Currently, AC stability and latency metrics are logged in
`UpdateResult()`. This CL moves the log to `NotifyChanged()`.

When the debouncing feature is disabled, these are mostly equivalent,
the former calls the latter with minimal computation in between. Though
the latter is additionally called when closing the popup, that didn't
matter since the metrics aren't logged when the popup is closed
regardless. So this change is mostly a no-op.

When debouncing is enabled, these differ, this change is no longer a
no-op. `UpdateResult()` occurs before the debouncing delay, while
`NotifyChanged()` occurs after. Since the metrics proxy the user-visible
stability and latency, they should capture the debouncing delay.

Adds '2' suffix to histogram names.

Since we're renaming the histogram anyways, takes the opportunity to
restrict the metrics to async requests. (i.e. excludes
`AutocompleteClassifier` requests). This was already done for some
metrics, but now it's done for all the metrics.

Bug: 1340291
Change-Id: I8a7a55003ca17b6a0ec75667cd3c7f835a9fbd9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3953073
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059549}