Track time spent in feed for Good Visits
Time is counted for periods where at least one content slice is 50% visible or covers 25% of the viewport. The existing implementation breaks up actions into segments that start when the layout changes due to scrolling or adding/removing Views; onPreDraw() seems to be called similarly and is a convenient place for measuring exposure and coverage, too.

Bug: b/243544984
Change-Id: I2d4d61bba67cde2de4e782cd53a2127603ff0dbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3949543
Reviewed-by: Dan H <harringtond@chromium.org>
Commit-Queue: Ian Wells <iwells@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059550}