Dictation: Create SandboxedPumpkinTagger from PumpkinData
This change does the work to unpack downloaded PumpkinData and
initialize Pumpkin in a sandboxed iframe. This is done in a sandbox
for security considerations, since we're required to call eval() and
override the fetch() method. Please see more details in the associated
design document [1]. The implementation is paired with a test file
that verifies the sandbox can be setup and used to parse speech.

This CL makes a few other changes, notably:
- Simplifies how PumpkinParseStrategy is created
- Removes logic around include/exclude pumpkin and instead only
unpacks compressed pumpkin files during tests.

With this change, Dictation can use Pumpkin for speech parsing. For
example, a user can say "Delete two characters" and Dictation will
fulfill the request; this was not possible with the previous
regex-based speech parsing.

[1] https://docs.google.com/document/d/1nZxo9rG0kXKbI4ee_dSudxFWhsciazHIN-CZGgkjBsE/edit?usp=sharing

AX-Relnotes: N/A
Bug: 1258190
Change-Id: Icc28db72d0bad19e89c49bdefe5a1ade74f2722e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3876464
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059551}