[Chromecast] Add or() operator to Observable.
This introduces a new way to combine two Observables.

The primary use case is to make one Observable out of two
expressions that return Observables with no regard as to
which one activates.

For example, in chromecast internal code, there are two ways
we can detect that the user may intend to resume playback of
a Cast app: the Android MediaSession and the Cast sender. With
Observable.or(), we can have an Observable for the media session
and an Observable for the Cast sender requests, and use .or() to
combine them into one Observable and subscribe a single observer
to handle both events the same way.

Bug: b/238697807 (enables a refactor for this bug)
Test: cast_base_junit_tests

Change-Id: Iea1673b62a71099c43e85854bb70c34f31b25b79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3957473
Commit-Queue: Yuchen Liu <yucliu@chromium.org>
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Auto-Submit: Simeon Anfinrud <sanfin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059748}