[VC]: add vc flags, connect to camera HAL server
This CL also removes some unused flag definitions.

This is a reland of https://crrev.com/c/3949174 with fixes to the
unit tests that were causing flakes. These fixes include a race
condition in ServerConnectionError where after disconnecting the
server it could try and reconnect before the disconnect handler
had been executed on the other side. Added in an observer to get
a signal when the handler had been executed and only reconnect
after that.

BUG=b:251337330
TEST=deploy to DUT, toggle flag on and off, see mojo call in logs

Change-Id: I300d2e6a997c111188eab41c457b06c705ce49e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3955002
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059794}