[hyphens] Implement layout part of `hyphenate-limit-chars`
This patch implements the layout part of the
`hyphenate-limit-chars` property, supporting the maximum
length of the word / before the hyphen / after the hyphen.

Because `kMinimum{Prefix,Suffix}Length` constants already
exist, theoretically supporting the prefix/suffix limits was
just to change the constants to the property values, plus
supporting the maximum limit of the word length. The constants
were not working as expected however, so this patch includes
minor changes to comparisons.

Bug: 924069
Change-Id: Id360e0e1e6cf1e99a76037bd976e9492953166f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3955123
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059824}