Reland of numeric quantity.
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/3549078
but due to a merge conflict, the reland could not be created in
gerrit.

Besides solving the merge conflict, no changes have been applied.

The problem that caused the revert was solved by rolling out new
parsing patterns. This was done by a commit to an external repo.

Bug: 807748
Change-Id: Id282d0f2b80a0561c74835d116aca4841bd85612
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3956876
Reviewed-by: Vidhan Jain <vidhanj@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Matthias Körber <koerber@google.com>
Cr-Commit-Position: refs/heads/main@{#1059834}