Destroy profile on browser close for Lacros secondary profiles
Part of the `kDestroyProfileOnBrowserClose` feature, the goal is to activate it on Lacros, but only on secondary profiles as a first step.

Added a new ProfileKeepAliveType `kLacrosMainProfile` to protect Lacros main profile from being destroyed on browser close.

Added a new browser test to test the new behavior on Lacros.

Change-Id: I36113b285e4ceb205d9f9743c99cefd56f433275
Bug: 1313848
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3941766
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1059849}