ForceSignin to use user profile instead of System Profile.
The profile picker used to use the system profile to host web contents for force signin and error dialogs; this CL changes to use the regular user profile instead.

The inline login handler now gets the information from the url created from the profile picker (now includes a `from_profile_picker=true` tag) to differentiate with other calls. It used to check if the profile used to render the web content was the System profile for that check.

This allowed to simply the code and the ProfilePicker API, and stop using the System profile is useful to stop creating KeyedServices for the irregular profiles when not really needed.

Profiles that failed to authenticate will now clear their FirstBrowserWindow keep alive to make sure that the loaded profile will be destroyed even if their first browser is not created.

Change-Id: Icd4c15f15719a87f4c58d5e017c0b3c0c4a68a06
Bug: 1374729
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3914411
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1059855}