[BrowserFragment] Add minimum setup to run instrumentation tests
The instrumentation tests run in the shell app and have access to the BF
public API.

This is based on the weblayer instrumentation tests.

For now, the tests run with the support apk, and the BF shell in local
mode.

Also re-enable the unit tests.

Change-Id: I06ec9d511029578e926fe80f96a9868fe336b581
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3950291
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Rayan Kanso <rayankans@chromium.org>
Reviewed-by: Susanne Westphal <swestphal@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059895}