[FileSystemURL] Plumb in the correct third-party StorageKey
CreateSnapshotFile() was incorrectly constructing its FileSystemURL with a first-party or origin to StorageKey conversion. To ensure that storage is partitioned correctly, it needed to use a third-party StorageKey retrieved from the renderer process.

This CL also removes the same stray TODO from the code as it has already been addressed.

Bug: 1372852
Change-Id: If68098ea6bc8ad8b3d962243da32660139279d43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3956158
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059940}