[Android] Don't listen to onPrimaryAccountChanged in SigininManagerImpl
Revises the sign-out flow in SigninManagerImpl. Before this CL,
SigninManagerImpl was listening to onPrimaryAccountChanged to proceed
to the data wiping during the sign-out flow. Unfortunately, this
approach is inherently racy, as both SigninManager and SyncService
are just observers of IdentityManager and there's no guarantee that
sync is going to be stopped before the data is wiped (at least for
managed accounts).

After this CL, wiping the data is just one more step in
SigninManagerImpl.signOut and revokeSyncConsent calls. This avoids
the possible race condition. As a bonus, this is also consistent
with how data wiping is implemented on other platforms.

Bug: 1344181
Change-Id: Ib15a0acd98b8a99d3be3c9694b1fb14ceebacc95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3875088
Reviewed-by: Tanmoy Mollik <triploblastic@chromium.org>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059941}