Use buildflag to determine Google branding in leak dialog.
Currently, Google branding is avoided iff the build is not
Chrome branded and the user is not syncing passwords. However,
this can only happen for non-branded developer builds with Google
API keys, as API keys are no longer supplied to non-Google builds.

This CL modifies the behavior to exclude this case: Google branding
will now be shown iff the build is Chrome branded. This allows some
simplications in the code and will allow repurposing the "IsSyncing"
property of a credential to indicate whether the credential is
actually syncing to a remote store (currently, it is false for account
store credentials and sync users with a passphrase).

Chrome branding in a developer build can be enabled by setting
is_chrome_branded=true.

Bug: 1368639
Change-Id: I22ff5e79975f4416e177aa0eba9c1e215791906f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3959934
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059985}